+++ demisto_sdk/commands/content_graph/objects/base_content.py
@@ -9,7 +9,6 @@ from typing import (
     Dict,
     List,
     Optional,
-    Sequence,
     Set,
     Type,
     cast,
@@ -193,26 +192,3 @@ class UnknownContent(BaseContent):
     @property
     def identifier(self):
         return self.object_id or self.name
-
-
-class FailedUploadException(RuntimeError):
-    def __init__(
-        self,
-        path: Path,
-        response_body: dict,
-        status_code: int,
-    ) -> None:
-        super().__init__(f"Failed uploading {path}, {status_code=}, {response_body=}")
-
-
-# TODO find a better place for these two
-class FailedUploadMultipleException(RuntimeError):
-    def __init__(self, failures: Sequence[FailedUploadException]) -> None:
-        self.failures = failures
-        super().__init__(self.failures)
-
-    def __str__(self) -> str:
-        return "\n".join(
-            "Failed uploading multiple content items:",
-            *[str(failure) for failure in self.failures],
-        )
